Move config-parsing logic to config module (no-op).
